+++ WarPlan/Locale.lua
@@ -12,11 +12,11 @@ V =
     }
     or C == "esMX" and { -- 17/20 (85%)
       "%d |4misión:misión; disponible", "%d |4misión:misiones; completadas", "%d |4misión:misiones; en progreso", "%d |4misión:misiones; restantes...", "%d |4grupo:grupos; pendientes...", "Coste base de misión: %s", "Recompensa de bonus", z, "Click para completar", "Completar todas",
-      "Completada hace %s ", "Termina en:", "Gran éxito", z, "Misión Fallida", "Historial de Misión", "Éxito de Misión", "No hay grupos adecuados", "Los resultados de las misiones completadas serán grabadas y mostradas aquí.", z,
+      "Completada hace %s", "Termina en:", "Gran éxito", z, "Misión Fallida", "Historial de Misión", "Éxito de Misión", "No hay grupos adecuados", "Los resultados de las misiones completadas serán grabadas y mostradas aquí.", z,
     }
     or C == "frFR" and { -- 20/20 (100%)
       "%d |4mission:missions; |4disponible:disponibles", "%d |4mission:missions; |4terminée:terminées", "%d |4mission:missions; en cours", "%d |4mission:missions; |4restante:restantes ...", "%d |4groupe:groupes; |4restant:restants ...", "Coût de la mission: %s", "Récompense Bonus", "Effacer la sélection en cours", "Cliquer pour terminer", "Tout Terminer",
-      "Terminé il y a %s", "Expire dans:", "Succès Total", "En Groupe:", "Mission Échouée", "Historique des Missions", "Mission Accomplie", "Aucun groupe adapté.", "L'historique des précédentes missions terminées sera enregistré et affiché ici.", "Envoyer en mission",
+      "Terminé il y a %s", "Expire dans:", "Grandiose Succès", "En Groupe:", "Mission Échouée", "Historique des Missions", "Mission Accomplie", "Aucun groupe adapté.", "L'historique des précédentes missions terminées sera enregistré et affiché ici.", "Envoyer en mission",
     }
     or C == "itIT" and { -- 20/20 (100%)
       "%d |4 missione disponibile:missioni disponibili;", "%d |4 missione completata:missioni completate;", "%d |4 missione:missioni; in corso", "%d |4 missione rimanente:missioni rimanenti; ...", "%d |4 gruppo rimanente:gruppi rimanenti; ...", "Costo missione base: %s", "Ricompensa del tiro bonus", "Cancella tutti i gruppi provvisori", "Clicca per completare", "Completa tutto",
@@ -27,8 +27,8 @@ V =
       "%s 전 완료됨", "만료:", "위대한 성공", "임시 그룹", "임무 실패", "임무 기록", "임무 성공", "생존 가능 그룹 없음", "완료한 임무 결과를 이곳에서 확인 가능", "임시 그룹 보내기",
     }
     or C == "ruRU" and { -- 20/20 (100%)
-      "%d |4миссия доступна:миссии доступны:миссий доступно; ", "%d |4миссия завершена:миссии завершены:миссий завершено;", "%d |4миссия:миссии:миссий; в процессе ", "%d |4миссия:миссии:миссий; осталось....", "%d |4группа осталась:группы осталось:групп осталось;...", "Стоимость миссии: %s", "Награда бонусного броска", "Очистить все пробные группы", "Нажмите, чтобы завершить", "Завершить все",
-      "Завершено %s назад", "Заканчивается через", "Грандиозный  успех", "В пробной группе:", "Миссия провалена", "История миссий ", "Миссия  выполнена", "Нет возможных групп", "Результаты выполненных миссий будут записаны и отображены здесь.", "Отправить пробные группы",
+      "%d |4миссия доступна:миссии доступны:миссий доступно;", "%d |4миссия завершена:миссии завершены:миссий завершено;", "%d |4миссия:миссии:миссий; в процессе", "%d |4миссия:миссии:миссий; осталось....", "%d |4группа осталась:группы осталось:групп осталось;...", "Стоимость миссии: %s", "Награда бонусного броска", "Очистить все пробные группы", "Нажмите, чтобы завершить", "Завершить все",
+      "Завершено %s назад", "Заканчивается через", "Грандиозный  успех", "В пробной группе:", "Миссия провалена", "История миссий", "Миссия  выполнена", "Нет возможных групп", "Результаты выполненных миссий будут записаны и отображены здесь.", "Отправить пробные группы",
     }
     or C == "zhCN" and { -- 20/20 (100%)
       "%d |4任务:任务; 可派遣", "%d |4任务:任务; 已完成", "%d |4任务:任务; 进行中", "%d |4任务:任务; 持续中...", "%d |4队伍:队伍; 持续中...", "花费：%s", "额外拾取", "清除所有预设队列", "点击完成", "一键完成",