+++ BetterIceFishing/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Better Ice Fishing
 
-## [100200.01-Release](https://github.com/fubaWoW/BetterIceFishing/tree/100200.01-Release) (2023-11-05)
-[Full Changelog](https://github.com/fubaWoW/BetterIceFishing/compare/100107.01-Release...100200.01-Release) [Previous Releases](https://github.com/fubaWoW/BetterIceFishing/releases)
+## [100205.02-Release](https://github.com/fubaWoW/BetterIceFishing/tree/100205.02-Release) (2024-01-26)
+[Full Changelog](https://github.com/fubaWoW/BetterIceFishing/compare/100205.01-Release...100205.02-Release) [Previous Releases](https://github.com/fubaWoW/BetterIceFishing/releases)
 
-- Toc Bump  
+- Toc Fix